more support for logins and better inline image handling

+++ Push/NotificationManager.m
@@ -274,7 +274,7 @@
     const unsigned char *buf = data.bytes;
     NSInteger i;
     for (i=0; i<data.length; ++i) {
-        [sbuf appendFormat:@"%02X", (NSUInteger)buf[i]];
+        [sbuf appendFormat:@"%02X", (unsigned int)buf[i]];
     }
 
     return sbuf;
